+++ sdk/harambe/html_converter/html_to_markdown.py
@@ -1,0 +1,4 @@
+from markdownify import MarkdownConverter
+
+class HTMLToMarkdownConverter(MarkdownConverter):
+    pass
