+++ sdk/test/test_e2e.py
@@ -496,6 +496,7 @@ async def test_capture_html_with_different_options(server, observer, harness):
     assert replaced_element in replaced_head_data["html"]
     assert "Replaced Text" in replaced_head_data["text"]
 
+
 @pytest.mark.parametrize("harness", [playwright_harness, soup_harness])
 async def test_capture_html_conversion_types(server, observer, harness):
     url = f"{server}/heading"
@@ -522,6 +523,7 @@ async def test_capture_html_conversion_types(server, observer, harness):
 
     # Text doesn't include markdown syntax
     assert observer.data[1]["text"].strip() == "Heading"
+
 
 @pytest.mark.parametrize("harness", [playwright_harness, soup_harness])
 async def test_capture_html_element_not_found(server, observer, harness):
